+++ decomp/ST.exe/functions/00692C10/decomp.c
@@ -21,7 +21,7 @@ FUN_00692c10(Global_sub_00692C10_param_1Enum param_1,int param_2,undefined4 *par
 {
   char cVar1;
   int iVar3;
-  uint uVar3;
+  int uVar3;
   int iVar7;
   uint *puVar4;
   int iVar5;
@@ -61,7 +61,7 @@ FUN_00692c10(Global_sub_00692C10_param_1Enum param_1,int param_2,undefined4 *par
         _DAT_00853cf0 = 0;
 
         uVar3 = thunk_FUN_006938c0(param_2);
-        if (DAT_00853d84 <= uVar3) {
+        if (DAT_00853d84 <= (uint)uVar3) {
 
           iVar7 = ReportDebugMessage("E:\\__titans\\Maps\\DefObj.cpp",0x1ba,0,uVar3,
                                      "%s","Out of Range Entourage Group Index ");
