+++ src/ST.exe/source/recovered/by_owner/TradePanelTy.cpp
@@ -12,7 +12,7 @@ void __thiscall st::fn_00551800(TradePanelTy *this)
 
 {
   int iVar1;
-  uint uVar2;
+  int uVar2;
   int iVar2;
   uint uVar3;
   uint uVar4;
@@ -57,7 +57,7 @@ LAB_0055194f:
 
         uVar2 = st::fn_004016E0(DAT_0080874d);
       }
-      if ((int)uVar2 < iVar2) goto LAB_0055198e;
+      if (uVar2 < iVar2) goto LAB_0055198e;
       iVar2 = 1;
     }
     else {
